+++ src/main/java/de/samply/reporter/script/ScriptEngine.java
@@ -1,10 +1,10 @@
 package de.samply.reporter.script;
 
 import de.samply.reporter.context.CellContext;
+import de.samply.reporter.context.CellStyleContext;
 import de.samply.reporter.context.Context;
 import de.samply.reporter.template.script.Script;
 import de.samply.reporter.template.script.ScriptFramework;
-import org.apache.poi.ss.usermodel.Workbook;
 
 public interface ScriptEngine {
 
@@ -12,6 +12,7 @@ public interface ScriptEngine {
 
   ScriptResult generateRawResult(Script script, Context context) throws ScriptEngineException;
 
-  CellContext generateCellContext(Script script, Workbook workbook) throws ScriptEngineException;
+  CellContext generateCellContext(Script script, CellStyleContext cellStyleContext, Context context)
+      throws ScriptEngineException;
 
 }