+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2023-05-30]
+## [1.0.0 - 2023-06-15]
 ### Added
 - Gitignore
 - Dockerignore
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Download Report
 - Remove temporal files
 - Report Status
+- Context and CellContext for Engines
+- Report examples
 
 
 ### Changed